requirements.txt
# FastAPI and web server
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20

# Image processing
Pillow==11.3.0

# AI/ML
google-generativeai==0.8.5

# Spotify API
spotipy==2.25.1

# HTTP requests
httpx==0.28.1
requests==2.32.5

# Data validation
pydantic==2.11.7

# Environment variables
python-dotenv==1.1.1

loguru==0.7.3